Moved ProviderName and MongoDBMajorVersion to lib.
Cleaned up some inconsistant import statements in atlas.py
