+++ tests/test_clusters.py
@@ -8,7 +8,8 @@ from os import environ, getenv
 from atlasapi.atlas import Atlas
 from atlasapi.lib import AtlasPeriods, AtlasUnits, AtlasGranularities
 from json import dumps
-from atlasapi.clusters import AtlasBasicReplicaSet, MongoDBMajorVersion as mdb_version, ClusterConfig
+from atlasapi.clusters import AtlasBasicReplicaSet, ClusterConfig
+from lib import MongoDBMajorVersion as mdb_version
 from atlasapi.clusters import ClusterConfig, ProviderSettings, ReplicationSpecs, InstanceSizeName
 from atlasapi.clusters import RegionConfig, AdvancedOptions, TLSProtocols
 from tests import BaseTests
